add invoke rate height lock transfer transaction

+++ wisdom-core/src/main/java/org/wisdom/contract/RateheightlockDefinition/Extract.java
@@ -8,5 +8,5 @@ import lombok.*;
 @Builder
 public class Extract {
     private long extractheight;
-    private int surplus;
+    private long surplus;
 }
